+++ Sources/Sentry/SentrySessionReplay.m
@@ -117,12 +117,28 @@ SentrySessionReplay ()
     @synchronized(self) {
         [_displayLink invalidate];
         _isRunning = NO;
+        [self prepareSegmentUntil:_dateProvider.date];
+    }
+}
+
+- (void)resume
+{
+    if (_reachedMaximumDuration == YES) {
+        return;
+    }
+    @synchronized(self) {
+        if (_isRunning) {
+            return;
+        }
+        _videoSegmentStart = nil;
+        [_displayLink linkWithTarget:self selector:@selector(newFrame:)];
+        _isRunning = YES;
     }
 }
 
 - (void)dealloc
 {
-    [self stop];
+    [_displayLink invalidate];
 }
 
 - (void)captureReplayForEvent:(SentryEvent *)event;
@@ -203,7 +219,6 @@ SentrySessionReplay ()
     if (_isFullSession &&
         [now timeIntervalSinceDate:_sessionStart] > _replayOptions.maximumDuration) {
         _reachedMaximumDuration = YES;
-        [self prepareSegmentUntil:now];
         [self stop];
         return;
     }