+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add a touch tracker for replay (#4041)
 - Add enableMetricKitRawPayload (#4044)
+- Resume session replay when app enters foreground (#4053)
 
 ### Fixes